Change mocking strategy to support mocking multiple calls
Manually compile a list of calls and then assert that this list
is accurate.
Only needed for sendto; recvfrom is always and only called with
the buffer size.
assert_called_with() only checks the most recent call.
assert_has_calls() allows for extra calls without throwing an error,
so if you forget to add an expectation your tests could pass without
actually verifying the appropriate conditions.
call_args_list works but it is cumbersome - it is a list of tuples.